Update project listing and statuses as of Dec 14, 2023
- 1 moved to Complete
- 1 moved to Deferred
- Add 1 missing description
- List 1 older project
- 2 new projects In Progress

As to why Project 021 was previously unlisted,
I am not sure, but maybe my reasoning was that
only I was involved, not P1M.ORG. But, now I am
thinking that I am a part of it, too, and some
of the projects listed earlier also had only myself
as a contributor out of initial P1M.ORG members.
So, decided to list Project 021, too.